Multi-Scenario Persistence - Unwind Persistence From Baseline Scenario
This work was done to unwind the assumption that there is only ever a single scenario (i.e. the Baseline) that we can hang all of our other data off of. Instead, the work here will allow additional features to be created in the context of a multiple scenario world.

Note: This PR does not add any new functionality. It simply lays the foundation for the mutli-scenario work as outlined by #168 which will additionally require more persistence changes for things like listing all scenarios, duplicating a scenario, deleting a scenario, etc.