Add placement option for Tooltip
With this commit, placement="right" looks same as before.

But placement="top" has design errors, am not sure where to fix them yet. Do in CSS or use some sort of "tethering" library that determines the position?

Will need to check placement="bottom" and placement="left" as well, if needed.

Not sure if possible to make placement optional in TypeScript, with "right" the default value? But limited to the available values from the Popper library.